backend/requirements.txt
Flask
Flask-CORS
flask-restful
flask-sqlalchemy
python-dotenv
pytesseract
Pillow
requests
opencv-python   
faiss-cpu
google-cloud-translate
google-cloud-vision
openai
jsonify